+++ spec/rubocop/cop/infl/soft_line_length_spec.rb
@@ -18,19 +18,22 @@ RSpec.describe RuboCop::Cop::Infl::SoftLineLength do
   # results in your head without worrying about rounding.
 
   percentage = 50 # `let` doesn't seem to interpolate in the `it` strings...
-  let(:hard_limit) { 30 }
-  let(:soft_limit) { 20 }
+  let(:directive_line) { '# rubocop:enable Foo/Bar' }
+  let(:short_line) { '#' * soft_limit }
+  let(:long_line) { '#' * hard_limit }
+  let(:over_long_line) { '#' * (hard_limit + 1) }
+
+  let(:hard_limit) { soft_limit + 10 }
+  let(:soft_limit) { directive_line.length + 10 }
+  let(:ignore_cop_directives) { false }
   let(:cop_config) do
     {
       'HardLimit' => hard_limit,
       'SoftLimit' => soft_limit,
-      'AllowedLongLinePercentage' => percentage
+      'AllowedLongLinePercentage' => percentage,
+      'IgnoreCopDirectives' => ignore_cop_directives
     }
   end
-
-  let(:short_line) { '#' * soft_limit }
-  let(:long_line) { '#' * hard_limit }
-  let(:over_long_line) { '#' * (hard_limit + 1) }
 
   it 'finds line longer than `HardLimit`' do
     inspect_source(cop, over_long_line)
@@ -63,6 +66,32 @@ RSpec.describe RuboCop::Cop::Infl::SoftLineLength do
 
       expected = num_long - (num_long + num_short) * percentage / 100.0
       expect(cop.offenses.size).to eq(expected)
+    end
+  end
+
+  context 'IgnoreCopDirectives' do
+    let(:source) do
+      [
+        directive_line, directive_line,
+        long_line, long_line, long_line, long_line,
+        short_line, short_line
+      ]
+    end
+
+    context 'true' do
+      let(:ignore_cop_directives) { true }
+
+      it 'reports an error (of 6 lines 4 are over the soft limit)' do
+        inspect_source(cop, source)
+        expect(cop.offenses.size).to eq(1)
+      end
+    end
+
+    context 'false' do
+      it 'reports no error (of 8 lines 4 are over the soft limit)' do
+        inspect_source(cop, source)
+        expect(cop.offenses).to be_empty
+      end
     end
   end
 end