+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## master (unreleased)
 
+* Add `IgnoreCopDirectives` (default `true`) to `Infl/SoftLineLength`
+
 ### Changes
 
 ## 0.0.1 (2017-02-18)